requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
flask-bcrypt
flask-migrate
gunicorn
pymysql
cryptography
wtforms-sqlalchemy
